requirements.txt
colorama==0.3.3
contextlib2==0.4.0
MacFSEvents==0.4
mock==1.0.1
nose==1.3.4
python-termstyle==0.1.10
PyYAML==3.11
requests==2.5.1
six==1.9.0
sniffer==0.3.4
vcrpy==1.2.0
wrapt==1.10.2
